refactor: refactor list_checkpoints
